+++ E_Expedisi_Express/Program.cs
@@ -1,7 +1,14 @@
+using E_Expedisi_Express.Data; // Pastikan untuk mengimpor namespace yang sesuai
+using Microsoft.EntityFrameworkCore;
+
 var builder = WebApplication.CreateBuilder(args);
 
 // Add services to the container.
 builder.Services.AddControllersWithViews();
+
+// Tambahkan konfigurasi DbContext
+builder.Services.AddDbContext<ExpedisiDbContext>(options =>
+    options.UseSqlServer(builder.Configuration.GetConnectionString("SqlConnectionString")));
 
 var app = builder.Build();
 
@@ -23,5 +30,11 @@ app.UseAuthorization();
 app.MapControllerRoute(
     name: "default",
     pattern: "{controller=Home}/{action=Index}/{id?}");
+
+// Pastikan menambahkan route untuk halaman login jika diperlukan
+app.MapControllerRoute(
+    name: "login",
+    pattern: "Account/Login",
+    defaults: new { controller = "Account", action = "Login" });
 
 app.Run();
